+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current Master
 
+- hotfix - disable token list
 - Added a deprecation warning for web3 https://github.com/ethereum/mist/releases/tag/v0.9.0
 
 ## 3.9.6 2017-8-09